Moving interval code to SqlIntervals
