+++ sql-client/src/main/scala/raw/client/sql/writers/TypedResultSetJsonWriter.scala
@@ -15,7 +15,7 @@ package raw.client.sql.writers
 import com.fasterxml.jackson.core.{JsonEncoding, JsonFactory, JsonParser}
 import com.fasterxml.jackson.databind.{JsonNode, ObjectMapper}
 import raw.client.api._
-import raw.client.sql.SqlTypesUtils.{intervalToString, parseInterval}
+import raw.client.sql.SqlIntervals.{intervalToString, parseInterval}
 import raw.utils.RecordFieldsNaming
 
 import java.io.{IOException, OutputStream}
